refactor(download): moves auth code to own file auth.go
